Reverted breaking changes | addressed review comment
